requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arandr==0.1.11
argcomplete==3.6.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.12.3
blinker==1.9.0
blivet==3.11.0
blivet-gui==2.6.0
boto3==1.37.16
botocore==1.37.16
Brlapi==0.8.5
Brotli==1.1.0
cachetools==5.5.2
certifi==2023.5.7
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
crypt_r==3.13.1
cryptography==44.0.3
cssselect==1.3.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==5.2.1
dnf==4.23.0
executing==2.2.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.18.0
Flask==3.1.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.9.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
initial-setup==0.3.101
ipython==9.2.0
ipython_pygments_lexers==1.1.1
iso639==0.1.4
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langtable==0.0.68
libcomps==0.1.21
libdnf==0.74.0
louis==3.28.0
lxml==5.2.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mutagen==1.47.0
narwhals==1.40.0
nftables==0.1
numpy==2.2.6
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.2.1
pipenv==2025.0.2
platformdirs==4.3.8
ply==3.11
productmd==1.43
prompt_toolkit==3.0.51
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.25.1
pycparser==2.20
pycryptodomex==3.21.0
pycups==2.0.4
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
pyenchant==3.2.2
pygdbmi==0.11.0.0
Pygments==2.19.1
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.58
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-linux-procfs==0.7.3
python-meh==0.52
pytz==2025.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
queuelib==1.8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.25.1
rpm==4.20.1
rsa==4.9.1
s3transfer==0.11.4
Scrapy==2.12.0
selenium==4.32.0
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sentry-sdk==2.17.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
service-identity==24.2.0
setools==4.5.1
setuptools==69.2.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.8.2
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.45.1
systemd-python==235
tenacity==9.1.2
tldextract==5.3.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
virtualenv==20.31.2
w3lib==2.3.1
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xkbregistry==0.3
yt-dlp==2025.2.19
zope.interface==7.2
